Verification JWT SERVICE

+++ verification-service/src/main/java/com/hivein/verificationservice/repository/SecureTokenRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SecureTokenRepository extends JpaRepository<SecureToken, Integer> {
+public interface SecureTokenRepository extends JpaRepository<SecureToken, Long> {
     Optional<SecureToken> findByEmail(String email);
 }
